Adding source for brew(1) man page
Breaking out the man page and the man page source in case the owner
wants to cherry-pick changes. This commit will have the meat of the
changes, with all the content required for generating the man page. If
it is decided to store it somewhere other than /usr/local/share/man then
the generate_manpage.rb script can be modified.